+++ packages/proto-signing/src/signing.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/naming-convention */
+import { assert } from "@cosmjs/utils";
 import { Coin } from "cosmjs-types/cosmos/base/v1beta1/coin";
 import { SignMode } from "cosmjs-types/cosmos/tx/signing/v1beta1/signing";
 import { AuthInfo, SignDoc, SignerInfo } from "cosmjs-types/cosmos/tx/v1beta1/tx";
@@ -38,6 +39,13 @@ export function makeAuthInfoBytes(
   feePayer: string | undefined,
   signMode = SignMode.SIGN_MODE_DIRECT,
 ): Uint8Array {
+  // Required arguments 4 and 5 were added in CosmJS 0.29. Use runtime checks to help our non-TS users.
+  assert(
+    feeGranter === undefined || typeof feeGranter === "string",
+    "feeGranter must be undefined or string",
+  );
+  assert(feePayer === undefined || typeof feePayer === "string", "feePayer must be undefined or string");
+
   const authInfo = {
     signerInfos: makeSignerInfos(signers, signMode),
     fee: {